requirements.txt
duckdb==0.10.2
pandas==2.2.2
scikit-learn==1.5.1
fastapi==0.111.0
uvicorn==0.30.1
joblib==1.4.2
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.40.1
pyyaml==6.0.1
pytest==8.3.2
typer==0.12.3
